Simplify Time struct to represent a single timeline, restore speed.

+++ src/apps/snacman/snacman/simulations/snacgame/SimulationControl.h
@@ -24,11 +24,10 @@ struct SimulationControl
     bool mPlaying = true;
     bool mStep = false;
     bool mSaveGameState = true;
-    float mSpeedRatio{1.f};
+    int mSpeedRatio{1};
     std::size_t mSaveStateIndex = 0;
     int mSelectedSaveState = -1;
-    std::array<std::optional<SaveState>, gNumberOfSavedStates>
-        mPreviousGameState;
+    std::array<std::optional<SaveState>, gNumberOfSavedStates> mPreviousGameState;
 
     void saveState(ent::State && aState,
                    std::chrono::microseconds aFrameDuration,
